+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All documentation about Filebeat can be found here.
 ### Added
 
 ### Improvements
-- All Godeps dependencies were updated to master on 2015-10-21 [#122](https://github.com/elastic/filebeat/pull/122)
+- All Godeps dependencies were updated to master on 2015-10-21 [#122]
+- Set default value for ignore_older config to 10 minutes. #164
 
 ### Deprecated